requirements.txt
# Requirements needed for development and for building releases.
# See BUILD.rst for details about the latter
invoke >= 0.20
rellu >= 0.7
twine >= 1.12
wheel
docutils
pygments >= 2.8
pydantic
build

dataclasses
dataclasses-json
PyMuPDF==1.22.3
fitz
pyinstaller
pycryptodome==3.18.0
robotframework-excellib==2.0.1
robotframework-pythonlibcore==2.2.1
robotframework-seleniumlibrary==5.1.3
robotframework-selenium2library==3.0.0
robotframework-requests
robotframework-sshlibrary
robotframework-excel
selenium==3.141.0
urllib3==1.26.4
pyyaml
requests
tinyaes==1.0.4
google-cloud-core==2.3.2
pypdfium2==4.15.0
types-ipaddress